// how-to/here-supertabs-starter/client/src/framework/shapes/version-shapes.ts
export interface VersionInfo {
	/**
	 * The version of the the instance of this platform if provided
	 */
	app?: string;

	/**
	 * The version of the code of this platform
	 */
	platformClient?: string;

	/**
	 * The version of the HERE Core UI Platform client code
	 */
	workspacePlatformClient?: string;

	/**
	 * The version of the workspace client code
	 */
	workspaceClient?: string;

	/**
	 * The version of the HERE Core UI Components this platform is currently connected to
	 */
	workspace?: string;

	/**
	 * The version of the notification center that this platform instance has connected to
	 */
	notificationCenter?: string;

	/**
	 * The version of the runtime this platform instance is running against
	 */
	runtime?: string;

	/**
	 * The version of the rvm that was used to launch this platform
	 */
	rvm?: string;
}